move linenoise to its own package
