+++ linenoise/linenoise-helper.c
@@ -4,8 +4,7 @@
 void linenoiseCompletionStub(const char *, linenoiseCompletions *);
 void linenoiseHintsStub(const char *, int *color, int *bold);
 
-
 void linenoise_setup() {
-	linenoiseSetCompletionCallback(linenoiseCompletionStub);
-	linenoiseSetHintsCallback(linenoiseHintsStub);
+  linenoiseSetCompletionCallback(linenoiseCompletionStub);
+  linenoiseSetHintsCallback(linenoiseHintsStub);
 }
